refactor(desktop): extract the workspace and delivery page bodies into hooks
`CodeWorkspaceBody` and `PullRequestsSurface` were single functions that
mixed several independent state machines. Each concern now lives in its
own hook, and the two composition roots only compose them.

Workspace (`src/code/workspace/`):
- `useWorkspaceSessions`: the agent list, `?task=` selection, start, fork,
  reap, and tab close.
- `useBrowserTabs`: browser titles, start pages, and closing the native
  webview when a tab or the page goes.
- `useTerminalTabs`: shell labels, ending a shell with its tab, the
  terminal chord.
- `useEditorTabs`: file reveal, quick open and new-tab asks, tab drag.

Delivery (`src/code/delivery/`):
- `usePullRequestQuery`: the list, paging, and adopted summaries.
- `usePullRequestDetail`: selection, the route-target fence, the detail
  cache.
- `usePullRequestKeyboardNav`: arrow keys and Escape on the open detail.

Filters already live in the delivery page, so there is no filter hook.

One behavior nudge: `selectConversation` focuses the conversation pane
from the latest layout rather than the render closure, so a fork that
resolves after a tab change no longer reverts that change.

Each hook has renderHook tests pinning the edges the split had to reason
about: shells end once per tab, browsers close on unmount, the terminal
chord jumps and returns, keyboard walks clamp at the top and page at the
bottom, a hand-picked row outranks a late route target, the detail cache
evicts oldest first, and a stale `?task=` is replaced rather than pushed.
